requirements_train.txt
# python 3.9
numpy==1.21.6
protobuf==3.20.1
pandas==1.3.5
scikit-learn==1.0.2
skl2onnx==1.13
tune-sklearn==0.4.3
ray[tune]==2.0.0
torch==1.12.1
onnxruntime==1.12.1
ConfigSpace==0.6.0
mlflow==1.29.0
matplotlib==3.5.3
boto3==1.24.83
pydantic==1.9.1
dvc==2.10.2
seaborn==0.12.1
sclblonnx==0.1.14
